app/requirements.txt
feedparser
requests
beautifulsoup4
trafilatura
psycopg[binary]
pyyaml
qdrant-client==1.14.1
jinja2
sentence-transformers
google-generativeai
schedule
duckduckgo-search>=6.1.1
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0